refactor: move Stop into .interval
We'll re-use the `stop` namespace for the stop interval calculation.
